backend/requirements.txt
# Core Dependencies
Flask==2.3.0
Flask-CORS==4.0.0
python-dotenv==1.0.0
gunicorn==22.0.0

# AI/ML Libraries
google-generativeai>=0.8.3
groq>=0.11.0
Pillow>=11.0.0
